+++ bili-api/src/main/kotlin/dev/aaa1115910/biliapi/http/entity/toview/ToViewData.kt
@@ -100,7 +100,7 @@ data class ToViewItem(
 
     @Serializable
     data class Stat(
-        val view: Int,
+        val view: Long,
         val danmaku: Int,
         val reply: Int,
         val favorite: Int,
