+++ player/tv/src/main/kotlin/dev/aaa1115910/bv/player/tv/controller/ControllerVideoInfo.kt
@@ -247,7 +247,7 @@ fun ControllerVideoInfoBottom(
     seekData: VideoPlayerSeekState,
     idleIcon: String,
     movingIcon: String,
-    play: Int,
+    play: Long,
     danmaku: Int,
     like: Int,
     coin: Int,
